% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demonstration_Roxygen.R
\docType{package}
\name{Demonstration}
\alias{Demonstration}
\title{Démonstration}
\description{
Ce package est une simple démonstration de création et d'exportation d'un package R.

Le package comporte les fonctions suivantes:

square();

random_sample();

random_sample_hist().
}
\details{
Pour obtenir plus d'aide avec les fonctions du packages, vous pouvez faire les commandes suivantes:

?square();

?random_sample();

?random_sample_hist().
}
\author{
Alexandre Millette \email{alexandre.millette@enap.ca}
}
