% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demonstration_Fonction.R
\name{random_sample_hist}
\alias{random_sample_hist}
\title{random_sample_hist}
\usage{
random_sample_plot(range = "", size = "")
}
\arguments{
\item{range}{numeric, sample range}

\item{size}{numeric, sample size (n)}
}
\description{
Cette fonction permet de réaliser un histogramme à partir d'un échantillonage aléatoire.
}
